+++ Services/Restaurants/Models/RestaurantDetailedModel.cs
@@ -18,9 +18,9 @@ namespace Services.Restaurants.Models
 
         public string City { get; set; }
 
-        public long GeoLatitude { get; set; }
+        public double GeoLatitude { get; set; }
 
-        public long GeoLongitude { get; set; }
+        public double GeoLongitude { get; set; }
 
         public MenuModel Menu { get; set; }
     }
